+++ src/hideAnimation.js
@@ -1,17 +1,10 @@
-import { isFalse } from './staticMethods.js'
 import { supportBrowserAnimationEventOfName_end } from './defaultParameters.js'
 import validateBrowserCompatiblityAnimationEvent from './compatiblity.js'
 import resetScroll from './resetScroll.js'
 
 export default function excuteHideAnimation (options, currentDialogNode) {
-    // ie8 是否支持 animation.
-    if(document.querySelector && document.addEventListener) {
-        // 兼容ie9
-        if(validateBrowserCompatiblityAnimationEvent(currentDialogNode, supportBrowserAnimationEventOfName_end) == undefined) {
-            currentDialogNode.style.display = 'none';
-            resetScroll(' codialog-show', false);
-            return null;
-        }
+    // 兼容 animation.
+    if( validateBrowserCompatiblityAnimationEvent(currentDialogNode, supportBrowserAnimationEventOfName_end) != undefined ) {
         // animation动画加载
         this.animate(options).delay(100).fadeOut('fadeOut',{
             type: 'end',
@@ -22,6 +15,7 @@ export default function excuteHideAnimation (options, currentDialogNode) {
         }).render();
     }
     else {
+        // ie9 不兼容 animation.
         currentDialogNode.style.display = 'none';
         resetScroll(' codialog-show', false);
     }