+++ src/showAnimation.js
@@ -6,14 +6,8 @@ import { supportBrowserAnimationEventOfName_end } from './defaultParameters.js'
 export default function excuteShowAnimation (options, currentDialogNode) {
     var resetDefaultAnimation = 'bounceIn';
 
-    // ie8 是否支持 animation.
-    if(document.querySelector && document.addEventListener) {
-        // 兼容ie9
-        if(validateBrowserCompatiblityAnimationEvent(currentDialogNode, supportBrowserAnimationEventOfName_end) == undefined) {
-            currentDialogNode.style.display = 'block';
-            resetScroll(' codialog-show', true);
-            return null;
-        }
+    // 兼容 animation.
+    if( validateBrowserCompatiblityAnimationEvent(currentDialogNode, supportBrowserAnimationEventOfName_end) != undefined ) {
         if(isFalse(this.hasAnimation)) resetDefaultAnimation = this.customAnimation || resetDefaultAnimation;
         // animation动画加载
         this.animate(options).delay(100)[resetDefaultAnimation](resetDefaultAnimation,{
@@ -26,6 +20,7 @@ export default function excuteShowAnimation (options, currentDialogNode) {
         .render();
     }
     else {
+        // ie9 不兼容 animation.
         currentDialogNode.style.display = 'block';
         resetScroll(' codialog-show', true);
     }